# src/swagger/calendar.yaml
openapi: 3.0.3
info:
  title: Calendar API
  version: 1.0.0
servers:
  - url: http://localhost:3000

tags:
  - name: Calendar

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiError:
      type: object
      properties:
        errorCode:
          type: string
          example: "C001"
        reason:
          type: string
          example: "연동 실패"
        data:
          nullable: true
          example: null

    ApiFail:
      type: object
      properties:
        resultType:
          type: string
          example: "FAIL"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          nullable: true
          example: null

    # ------------- Google Connect -------------
    GoogleConnectRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: "4/0AeaYSH..."

    ApiSuccessMessage:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          type: object
          properties:
            message:
              type: string
              example: "연동 성공"

    # ------------- Google Status -------------
    GoogleStatusSuccess:
      type: object
      properties:
        isConnected:
          type: boolean
          example: true

    ApiGoogleStatus:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          $ref: "#/components/schemas/GoogleStatusSuccess"

    # ------------- Google Events Preview -------------
    GooglePreviewEvent:
      type: object
      properties:
        id:
          type: string
          example: "google_evt_abc"
        title:
          type: string
          example: "팀 미팅"

    ApiGoogleEventsPreview:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          type: object
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/GooglePreviewEvent"

    # ------------- Google Sync -------------
    GoogleSyncEventInput:
      type: object
      required: [id, summary, start, end]
      properties:
        id:
          type: string
        summary:
          type: string
        start:
          type: object
          additionalProperties: true
        end:
          type: object
          additionalProperties: true

    GoogleSyncRequest:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/GoogleSyncEventInput"

    ApiGoogleSync:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          type: object
          properties:
            savedCount:
              type: integer
              example: 5

    # ------------- Calendar Month -------------
    CalendarMonthPreviewItem:
      type: object
      properties:
        itemId:
          type: string
          example: "10"
        title:
          type: string
          example: "과제하기"
        isFixed:
          type: boolean
          example: false
        type:
          type: string
          example: "MANUAL"
        # ❌ 배포 서버 DB 컬럼 부재로 eventColor 제거

    CalendarMonthDay:
      type: object
      properties:
        date:
          type: string
          example: "2026-02-11"
        hasItems:
          type: boolean
          example: true
        itemCount:
          type: integer
          example: 4
        schedulesPreview:
          type: array
          items:
            $ref: "#/components/schemas/CalendarMonthPreviewItem"
        moreCount:
          type: integer
          example: 1

    CalendarMonthSuccess:
      type: object
      properties:
        year:
          type: integer
          example: 2026
        month:
          type: integer
          example: 2
        days:
          type: array
          items:
            $ref: "#/components/schemas/CalendarMonthDay"

    ApiCalendarMonth:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          $ref: "#/components/schemas/CalendarMonthSuccess"

    # ------------- Calendar Day -------------
    CalendarDayItem:
      type: object
      properties:
        itemId:
          type: string
          example: "10"
        title:
          type: string
          example: "과제하기"
        startAt:
          type: string
          example: "2026-02-11T18:00:00+09:00"
        endAt:
          type: string
          example: "2026-02-11T20:00:00+09:00"
        isFixed:
          type: boolean
          example: false
        type:
          type: string
          example: "MANUAL"
        # ❌ 배포 서버 DB 컬럼 부재로 eventColor, recurrenceRule 제거

    CalendarDaySuccess:
      type: object
      properties:
        date:
          type: string
          example: "2026-02-11"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CalendarDayItem"

    ApiCalendarDay:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          $ref: "#/components/schemas/CalendarDaySuccess"

    # ------------- Manual Create/Update -------------
    ManualCreateRequest:
      type: object
      required: [title, startAt, endAt]
      properties:
        title:
          type: string
          example: "과제 제출"
        startAt:
          type: string
          example: "2026-02-11T18:00:00+09:00"
        endAt:
          type: string
          example: "2026-02-11T20:00:00+09:00"
        type:
          type: string
          description: "MANUAL | FIXED"
          example: "MANUAL"
        # ❌ eventColor, recurrenceRule 제거

    ManualUpdateRequest:
      type: object
      properties:
        title:
          type: string
        startAt:
          type: string
        endAt:
          type: string
        type:
          type: string
        # ❌ eventColor, recurrenceRule 제거

    ManualEventResponse:
      type: object
      properties:
        id:
          type: integer
          example: 10
        title:
          type: string
        startAt:
          type: string
        endAt:
          type: string
        type:
          type: string
          example: "MANUAL"
        status:
          type: string
          example: "TODO"
        googleEventId:
          nullable: true
          example: null
        # ❌ eventColor, recurrenceRule 제거

    ApiManualEvent:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          $ref: "#/components/schemas/ManualEventResponse"

    ApiManualDelete:
      type: object
      properties:
        resultType:
          type: string
          example: "SUCCESS"
        error:
          nullable: true
          example: null
        success:
          type: object
          properties:
            message:
              type: string
              example: "삭제되었습니다."

security:
  - bearerAuth: []

paths:
  /api/integrations/google-calendar/connect:
    post:
      summary: 구글 캘린더 연동 (Auth Code 전달)
      tags: [Calendar]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleConnectRequest"
      responses:
        "200":
          description: 연동 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccessMessage"

  /api/integrations/google-calendar/status:
    get:
      summary: 연동 상태 조회
      tags: [Calendar]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiGoogleStatus"

  /api/integrations/google-calendar/events:
    get:
      summary: 구글 캘린더 일정 목록 조회 (미리보기)
      tags: [Calendar]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiGoogleEventsPreview"

  /api/integrations/google-calendar/sync:
    post:
      summary: 구글 캘린더 일정 선택 저장 (동기화)
      tags: [Calendar]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSyncRequest"
      responses:
        "200":
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiGoogleSync"

  /api/calendar/month:
    get:
      summary: 월간 일정 조회 (캘린더 뷰)
      tags: [Calendar]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: year
          required: true
          schema:
            type: integer
        - in: query
          name: month
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCalendarMonth"

  /api/calendar/day:
    get:
      summary: 일간 일정 상세 조회 (리스트 뷰)
      tags: [Calendar]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            example: "2026-02-11"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCalendarDay"

  /api/calendar/event:
    post:
      summary: 직접 일정 추가 (Manual Create)
      tags: [Calendar]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualCreateRequest"
      responses:
        "201":
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiManualEvent"

  /api/calendar/event/{id}:
    patch:
      summary: 직접 일정 수정 (Manual Update)
      tags: [Calendar]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualUpdateRequest"
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiManualEvent"

    delete:
      summary: 직접 일정 삭제 (Manual Delete)
      tags: [Calendar]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiManualDelete"